backend/requirements.txt
bcrypt==4.2.0
blinker==1.8.2
boto3==1.35.39
botocore==1.35.39
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==22.0.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
packaging==24.0
pluggy==1.5.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.9.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.3
setuptools==69.5.1
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.11.0
urllib3==2.2.3
Werkzeug==3.0.3
wheel==0.43.0
